Add latest code for Token Action Maker
I'm changing the documentation of Token Action Maker on Confluence so that it is clear that Prod Wiz holds the code and that it does not need to be separately installed. I have also added buttons for the most commonly used commands to the Token Helper menu. Finally, I have incremented the version number to 0.8.3

In the future, whenever I update the Tam Code, I will submit a PR here.